Added XPathRecord class.
This is an immutable record, because the way to modify an XML document
is not obvious. For instance there is no obious way to add a new
field, (should it be attached under the root, should we create a new
XML node that is the parent of all new attachments, etc).

Therefore the Record interface represents an immutable record, that
can only be queried, but not changed. Also this interface provides a
way (the method makeMutable) to get a mutable record with the same
content.

Finally the MutableRecord interface, contains all the former methods
of the interface Record that used to change the record.

XPathRecord is now a Record, and MapRecord is a MutableRecord.