Fixed DSpaceOutputGenerator bug
The DSpaceOutputGenerator was producing empty .xml file if there was a
field with a namespace in the field map that did not contain any
data. This is resolved by moving the production of the file JSON
representation in another (private) method, and writing the final
representation only if there exist data for the respective namespace.